+++ app/src/main/java/com/sefford/artdrian/common/language/files/SizeUnit.kt
@@ -1,4 +1,4 @@
-package com.sefford.artdrian.common.language.units
+package com.sefford.artdrian.common.language.files
 
 import com.sefford.artdrian.common.language.isSingleUnit
 import com.sefford.artdrian.common.language.toString
