+++ app/src/main/java/com/sefford/artdrian/downloads/domain/model/Measured.kt
@@ -1,6 +1,6 @@
 package com.sefford.artdrian.downloads.domain.model
 
-import com.sefford.artdrian.common.language.units.Size
+import com.sefford.artdrian.common.language.files.Size
 
 interface Measured {
