+++ app/src/test/java/com/sefford/artdrian/downloads/store/LoadedDownloadsStateTest.kt
@@ -1,7 +1,7 @@
 package com.sefford.artdrian.downloads.store
 
 import com.sefford.artdrian.common.data.DataError
-import com.sefford.artdrian.common.language.units.Size.Companion.bytes
+import com.sefford.artdrian.common.language.files.Size.Companion.bytes
 import com.sefford.artdrian.downloads.domain.model.Download
 import com.sefford.artdrian.test.mothers.DownloadsMother
 import io.kotest.matchers.collections.shouldHaveSize
