+++ app/src/test/java/com/sefford/artdrian/test/assertions/Size.kt
@@ -1,6 +1,6 @@
 package com.sefford.artdrian.test.assertions
 
-import com.sefford.artdrian.common.language.units.Size
+import com.sefford.artdrian.common.language.files.Size
 import io.kotest.matchers.longs.shouldBeExactly
